[report] BigDecimal处理

+++ perfect-service/src/main/java/com/perfect/entity/RegionReportEntity.java
@@ -3,6 +3,8 @@ package com.perfect.entity;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Field;
 
+import java.math.BigDecimal;
+
 import static com.perfect.mongodb.utils.EntityConstants.*;
 
 /**
@@ -41,13 +43,13 @@ public class RegionReportEntity extends AccountIdEntity {
     private Double pcCtr;     //PC点击率=点击次数/展现次数
 
     @Field(value = "pccost")
-    private Double pcCost;        //PC消费
+    private BigDecimal pcCost;        //PC消费
 
     @Field(value = "pccpc")
-    private Double pcCpc;     //PC平均点击价格=消费/点击次数
+    private BigDecimal pcCpc;     //PC平均点击价格=消费/点击次数
 
     @Field(value = "pccpm")
-    private Double pcCpm;       //PC千次展现消费
+    private BigDecimal pcCpm;       //PC千次展现消费
 
     @Field(value = "pccs")
     private Double pcConversion;      //PC转化
@@ -65,13 +67,13 @@ public class RegionReportEntity extends AccountIdEntity {
     private Double mobileCtr;
 
     @Field(value = "mcost")
-    private Double mobileCost;
+    private BigDecimal mobileCost;
 
     @Field(value = "mcpc")
-    private Double mobileCpc;
+    private BigDecimal mobileCpc;
 
     @Field(value = "mcpm")
-    private Double mobileCpm;
+    private BigDecimal mobileCpm;
 
     @Field(value = "mcs")
     private Double mobileConversion;
@@ -159,27 +161,27 @@ public class RegionReportEntity extends AccountIdEntity {
         this.pcCtr = pcCtr;
     }
 
-    public Double getPcCost() {
+    public BigDecimal getPcCost() {
         return pcCost;
     }
 
-    public void setPcCost(Double pcCost) {
+    public void setPcCost(BigDecimal pcCost) {
         this.pcCost = pcCost;
     }
 
-    public Double getPcCpc() {
+    public BigDecimal getPcCpc() {
         return pcCpc;
     }
 
-    public void setPcCpc(Double pcCpc) {
+    public void setPcCpc(BigDecimal pcCpc) {
         this.pcCpc = pcCpc;
     }
 
-    public Double getPcCpm() {
+    public BigDecimal getPcCpm() {
         return pcCpm;
     }
 
-    public void setPcCpm(Double pcCpm) {
+    public void setPcCpm(BigDecimal pcCpm) {
         this.pcCpm = pcCpm;
     }
 
@@ -223,27 +225,27 @@ public class RegionReportEntity extends AccountIdEntity {
         this.mobileCtr = mobileCtr;
     }
 
-    public Double getMobileCost() {
+    public BigDecimal getMobileCost() {
         return mobileCost;
     }
 
-    public void setMobileCost(Double mobileCost) {
+    public void setMobileCost(BigDecimal mobileCost) {
         this.mobileCost = mobileCost;
     }
 
-    public Double getMobileCpc() {
+    public BigDecimal getMobileCpc() {
         return mobileCpc;
     }
 
-    public void setMobileCpc(Double mobileCpc) {
+    public void setMobileCpc(BigDecimal mobileCpc) {
         this.mobileCpc = mobileCpc;
     }
 
-    public Double getMobileCpm() {
+    public BigDecimal getMobileCpm() {
         return mobileCpm;
     }
 
-    public void setMobileCpm(Double mobileCpm) {
+    public void setMobileCpm(BigDecimal mobileCpm) {
         this.mobileCpm = mobileCpm;
     }
